wilayah, bahan makanan, geolokasi, result diagnosa, jadwal makan. obat

+++ database/migrations/2024_05_19_143017_create_certainty_factors_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreign('id_penyakit')->references('id')->on('penyakits')->onDelete('cascade');
             $table->unsignedBigInteger('id_gejala')->nullable();
             $table->foreign('id_gejala')->references('id')->on('gejalas')->onDelete('cascade');
-            $table->float('certainty_factor');
+            $table->double('certainty_factor', 8, 1);
             $table->timestamps();
         });
     }
